feat!: require ACME success
Before configuring Traefik with a new defaultGeneratedCert, try to
obtain it for a temporary router. Only when the requested certificate is
in acme.json, apply the new configuration safely.

BREAKING CHANGE: the set-certificate becomes synchronous. It returns
when a certificate has been obtained, or there is an error. If an error
occurs, it is printed to stderr.